Remove duplicated earned badge service
