+++ packages/server/src/app/badge/routers/earned-badge.router.js
@@ -2,8 +2,8 @@ import express from 'express';
 import { checkSchema, matchedData } from 'express-validator';
 import { jwtAuthHeaderValidator } from '../../auth/index.js';
 import { entityIdValidator } from '../../core/middleware/index.js';
-import { earnedBadgeController } from '../controllers/earned_badge.controller.js';
-import { earnedBadgeSchema, newEarnedBadgeSchema } from '../validators/earned_badge.validator.js';
+import { earnedBadgeController } from '../controllers/earned-badge.controller.js';
+import { earnedBadgeSchema, newEarnedBadgeSchema } from '../validators/earned-badge.validator.js';
 
 export const earnedBadgeRouter = express.Router();
