+++ packages/server/src/app/badge/validators/earned-badge.validator.js
